requirements.txt
Flask==1.0.2
pandas==0.23.4
numpy==1.15.1
requests==2.20.0
mysql-connector==2.1.6
mysqlclient==1.3.12
scikit_learn==0.20.3
gunicorn==19.7.1
python-dateutil==2.6.1